+++ CHANGELOG.md
@@ -116,6 +116,7 @@
 ## Features
 
 - [Issue #51](https://github.com/influxdb/influxdb/issues/51). Implement first and last aggregates
+- [Issue #35](https://github.com/influxdb/influxdb/issues/35). Support table aliases in Join Queries
 
 ## Bugfixes